+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
     - sliders: ignore key order in `range`
+    - sliders: handle mousewheel increment when starting from a value between two steps with `steps` property defined
 
 - widgets
     - encoder: expose angle in script (as `locals.angle`)