+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - sliders: ignore key order in `range`
     - sliders: handle mousewheel increment when starting from a value between two steps with `steps` property defined
     - inspector: allow scrolling in the help modal
+    - console: fix command history behavior and increase history size
 
 - widgets
     - encoder: expose angle in script (as `locals.angle`)